move the logic of preprocessing to the separate class 'PreprocessHandler' in 'syngen/ml/preprocess'
