--- src/syngen/__init__.py
@@ -1,1 +1,0 @@
-from syngen.train import preprocess_data  # noqa: F401
